[B] refactor : Auth : 패키지 이름 auth로 수정 #31
