+++ backend/src/main/java/ssu/groupstudy/domain/auth/service/CustomUserDetailService.java
@@ -1,11 +1,11 @@
-package ssu.groupstudy.domain.login.service;
+package ssu.groupstudy.domain.auth.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
-import ssu.groupstudy.domain.login.security.CustomUserDetails;
+import ssu.groupstudy.domain.auth.security.CustomUserDetails;
 import ssu.groupstudy.domain.user.domain.User;
 import ssu.groupstudy.domain.user.exception.UserNotFoundException;
 import ssu.groupstudy.domain.user.repository.UserRepository;
